[Fix] fix output type format

+++ LABs/lab1/np_sample_code/udp/lab1/client.c
@@ -100,7 +100,7 @@ int main(int argc, char *argv[])
     clock_t end_time = clock();
     double time_spent = (double)(end_time - start_time) / CLOCKS_PER_SEC;
     printf("\n[+] Total data received:%ld\n[+] File size:%ld\n", recv_bytes, file_size);
-    printf("[+] Packet Loss Rate(loss data size/total size):%lf\n", (file_size-recv_bytes)/file_size);
+    printf("[+] Packet Loss Rate(loss data size/total size):%lf\n", (((double)(file_size-recv_bytes))/(double)file_size));
     printf("[+] Total tran time: %lf ms\n", (time_spent*1000));
     printf("[+] FIle Size:%ld MB\n", (file_size/1024)/1024);
     sendto(sockfd, "OK", sizeof("OK"), sendrecvflag, (struct sockaddr*)&addr_con, addrlen);
